+++ src/ClientBuilder.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Sentry;
 
-use Http\Client\Common\Plugin;
+use Http\Client\Common\Plugin as PluginInterface;
 use Http\Client\Common\Plugin\AuthenticationPlugin;
 use Http\Client\Common\Plugin\BaseUriPlugin;
 use Http\Client\Common\Plugin\DecoderPlugin;
@@ -19,9 +19,9 @@ use Http\Discovery\HttpAsyncClientDiscovery;
 use Http\Discovery\MessageFactoryDiscovery;
 use Http\Discovery\StreamFactoryDiscovery;
 use Http\Discovery\UriFactoryDiscovery;
-use Http\Message\MessageFactory;
+use Http\Message\MessageFactory as MessageFactoryInterface;
 use Http\Message\StreamFactory as StreamFactoryInterface;
-use Http\Message\UriFactory;
+use Http\Message\UriFactory as UriFactoryInterface;
 use Jean85\PrettyVersions;
 use Sentry\HttpClient\Authentication\SentryAuthentication;
 use Sentry\HttpClient\Plugin\GzipEncoderPlugin;
@@ -29,6 +29,7 @@ use Sentry\Integration\ErrorListenerIntegration;
 use Sentry\Integration\ExceptionListenerIntegration;
 use Sentry\Integration\FatalErrorListenerIntegration;
 use Sentry\Integration\RequestIntegration;
+use Sentry\Integration\TransactionIntegration;
 use Sentry\Serializer\RepresentationSerializer;
 use Sentry\Serializer\RepresentationSerializerInterface;
 use Sentry\Serializer\Serializer;
@@ -50,7 +51,7 @@ final class ClientBuilder implements ClientBuilderInterface
     private $options;
 
     /**
-     * @var UriFactory|null The PSR-7 URI factory
+     * @var UriFactoryInterface|null The PSR-7 URI factory
      */
     private $uriFactory;
 
@@ -60,12 +61,12 @@ final class ClientBuilder implements ClientBuilderInterface
     private $streamFactory;
 
     /**
-     * @var MessageFactory|null The PSR-7 message factory
+     * @var MessageFactoryInterface|null The PSR-7 message factory
      */
     private $messageFactory;
 
     /**
-     * @var TransportInterface The transport
+     * @var TransportInterface|null The transport
      */
     private $transport;
 
@@ -75,17 +76,17 @@ final class ClientBuilder implements ClientBuilderInterface
     private $httpClient;
 
     /**
-     * @var Plugin[] The list of Httplug plugins
+     * @var PluginInterface[] The list of Httplug plugins
      */
     private $httpClientPlugins = [];
 
     /**
-     * @var SerializerInterface The serializer to be injected in the client
+     * @var SerializerInterface|null The serializer to be injected in the client
      */
     private $serializer;
 
     /**
-     * @var RepresentationSerializerInterface The representation serializer to be injected in the client
+     * @var RepresentationSerializerInterface|null The representation serializer to be injected in the client
      */
     private $representationSerializer;
 
@@ -107,6 +108,7 @@ final class ClientBuilder implements ClientBuilderInterface
     public function __construct(Options $options = null)
     {
         $this->options = $options ?? new Options();
+        $this->sdkVersion = PrettyVersions::getVersion('sentry/sentry')->getPrettyVersion();
 
         if ($this->options->hasDefaultIntegrations()) {
             $this->options->setIntegrations(array_merge([
@@ -114,6 +116,7 @@ final class ClientBuilder implements ClientBuilderInterface
                 new ErrorListenerIntegration(null, false),
                 new FatalErrorListenerIntegration(),
                 new RequestIntegration(),
+                new TransactionIntegration(),
             ], $this->options->getIntegrations()));
         }
     }
@@ -137,7 +140,7 @@ final class ClientBuilder implements ClientBuilderInterface
     /**
      * {@inheritdoc}
      */
-    public function setUriFactory(UriFactory $uriFactory): ClientBuilderInterface
+    public function setUriFactory(UriFactoryInterface $uriFactory): ClientBuilderInterface
     {
         $this->uriFactory = $uriFactory;
 
@@ -147,7 +150,7 @@ final class ClientBuilder implements ClientBuilderInterface
     /**
      * {@inheritdoc}
      */
-    public function setMessageFactory(MessageFactory $messageFactory): ClientBuilderInterface
+    public function setMessageFactory(MessageFactoryInterface $messageFactory): ClientBuilderInterface
     {
         $this->messageFactory = $messageFactory;
 
@@ -177,7 +180,7 @@ final class ClientBuilder implements ClientBuilderInterface
     /**
      * {@inheritdoc}
      */
-    public function addHttpClientPlugin(Plugin $plugin): ClientBuilderInterface
+    public function addHttpClientPlugin(PluginInterface $plugin): ClientBuilderInterface
     {
         $this->httpClientPlugins[] = $plugin;
 
@@ -231,20 +234,6 @@ final class ClientBuilder implements ClientBuilderInterface
     }
 
     /**
-     * Gets the SDK version to be passed onto {@see Event} and HTTP User-Agent header.
-     *
-     * @return string
-     */
-    private function getSdkVersion(): string
-    {
-        if (null === $this->sdkVersion) {
-            $this->setSdkVersionByPackageName('sentry/sentry');
-        }
-
-        return $this->sdkVersion;
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function setSdkVersion(string $sdkVersion): ClientBuilderInterface
@@ -260,9 +249,13 @@ final class ClientBuilder implements ClientBuilderInterface
      * @param string $packageName The package name that will be used to get the version from (i.e. "sentry/sentry")
      *
      * @return $this
+     *
+     * @deprecated since version 2.2, to be removed in 3.0
      */
     public function setSdkVersionByPackageName(string $packageName): ClientBuilderInterface
     {
+        @trigger_error(sprintf('Method %s() is deprecated since version 2.2 and will be removed in 3.0.', __METHOD__), E_USER_DEPRECATED);
+
         $this->sdkVersion = PrettyVersions::getVersion($packageName)->getPrettyVersion();
 
         return $this;
@@ -301,8 +294,8 @@ final class ClientBuilder implements ClientBuilderInterface
             $this->addHttpClientPlugin(new BaseUriPlugin($this->uriFactory->createUri($this->options->getDsn())));
         }
 
-        $this->addHttpClientPlugin(new HeaderSetPlugin(['User-Agent' => $this->sdkIdentifier . '/' . $this->getSdkVersion()]));
-        $this->addHttpClientPlugin(new AuthenticationPlugin(new SentryAuthentication($this->options, $this->sdkIdentifier, $this->getSdkVersion())));
+        $this->addHttpClientPlugin(new HeaderSetPlugin(['User-Agent' => $this->sdkIdentifier . '/' . $this->sdkVersion]));
+        $this->addHttpClientPlugin(new AuthenticationPlugin(new SentryAuthentication($this->options, $this->sdkIdentifier, $this->sdkVersion)));
 
         if ($this->options->isCompressionEnabled()) {
             $this->addHttpClientPlugin(new GzipEncoderPlugin($this->streamFactory));
@@ -343,16 +336,14 @@ final class ClientBuilder implements ClientBuilderInterface
                 throw new \RuntimeException('The `http_proxy` option requires the `php-http/curl-client` package to be installed.');
             }
 
+            /** @psalm-suppress InvalidPropertyAssignmentValue */
             $this->httpClient = new HttpCurlClient(null, null, [
                 CURLOPT_PROXY => $this->options->getHttpProxy(),
             ]);
         }
 
+        /** @psalm-suppress PossiblyInvalidPropertyAssignmentValue */
         $this->httpClient = $this->httpClient ?? HttpAsyncClientDiscovery::find();
-
-        if (null === $this->messageFactory) {
-            throw new \RuntimeException('The PSR-7 message factory must be set.');
-        }
 
         return new HttpTransport($this->options, $this->createHttpClientInstance(), $this->messageFactory);
     }
@@ -367,6 +358,6 @@ final class ClientBuilder implements ClientBuilderInterface
         $this->serializer = $this->serializer ?? new Serializer($this->options);
         $this->representationSerializer = $this->representationSerializer ?? new RepresentationSerializer($this->options);
 
-        return new EventFactory($this->serializer, $this->representationSerializer, $this->options, $this->sdkIdentifier, $this->getSdkVersion());
+        return new EventFactory($this->serializer, $this->representationSerializer, $this->options, $this->sdkIdentifier, $this->sdkVersion);
     }
 }