+++ src/Context/TagsContext.php
@@ -21,13 +21,7 @@ class TagsContext extends Context
             throw new \InvalidArgumentException('The tags context does not allow recursive merging of its data.');
         }
 
-        foreach ($data as $value) {
-            if (!\is_string($value)) {
-                throw new \InvalidArgumentException('The $data argument must contains a simple array of string values.');
-            }
-        }
-
-        parent::merge($data);
+        parent::merge(self::sanitizeData($data));
     }
 
     /**
@@ -35,13 +29,7 @@ class TagsContext extends Context
      */
     public function setData(array $data): void
     {
-        foreach ($data as $value) {
-            if (!\is_string($value)) {
-                throw new \InvalidArgumentException('The $data argument must contains a simple array of string values.');
-            }
-        }
-
-        parent::setData($data);
+        parent::setData(self::sanitizeData($data));
     }
 
     /**
@@ -49,13 +37,7 @@ class TagsContext extends Context
      */
     public function replaceData(array $data): void
     {
-        foreach ($data as $value) {
-            if (!\is_string($value)) {
-                throw new \InvalidArgumentException('The $data argument must contains a simple array of string values.');
-            }
-        }
-
-        parent::replaceData($data);
+        parent::replaceData(self::sanitizeData($data));
     }
 
     /**
@@ -63,10 +45,39 @@ class TagsContext extends Context
      */
     public function offsetSet($offset, $value): void
     {
+        if (is_numeric($value)) {
+            $value = (string) $value;
+        }
+
         if (!\is_string($value)) {
             throw new \InvalidArgumentException('The $value argument must be a string.');
         }
 
         parent::offsetSet($offset, $value);
+    }
+
+    /**
+     * Sanitizes the given data by converting numeric values to strings.
+     *
+     * @param array $data The data to sanitize
+     *
+     * @return array
+     *
+     * @throws \InvalidArgumentException If any of the values of the input data
+     *                                   is not a number or a string
+     */
+    private static function sanitizeData(array $data): array
+    {
+        foreach ($data as &$value) {
+            if (is_numeric($value)) {
+                $value = (string) $value;
+            }
+
+            if (!\is_string($value)) {
+                throw new \InvalidArgumentException('The $data argument must contains a simple array of string values.');
+            }
+        }
+
+        return $data;
     }
 }