+++ src/Integration/ErrorListenerIntegration.php
@@ -8,7 +8,7 @@ use Sentry\ErrorHandler;
 use Sentry\Exception\FatalErrorException;
 use Sentry\Exception\SilencedErrorException;
 use Sentry\Options;
-use Sentry\State\Hub;
+use Sentry\SentrySdk;
 
 /**
  * This integration hooks into the global error handlers and emits events to
@@ -51,13 +51,14 @@ final class ErrorListenerIntegration implements IntegrationInterface
      */
     public function setupOnce(): void
     {
+        /** @psalm-suppress DeprecatedMethod */
         $errorHandler = ErrorHandler::registerOnce(ErrorHandler::DEFAULT_RESERVED_MEMORY_SIZE, false);
         $errorHandler->addErrorHandlerListener(function (\ErrorException $exception): void {
             if (!$this->handleFatalErrors && $exception instanceof FatalErrorException) {
                 return;
             }
 
-            $currentHub = Hub::getCurrent();
+            $currentHub = SentrySdk::getCurrentHub();
             $integration = $currentHub->getIntegration(self::class);
             $client = $currentHub->getClient();